requirements.txt
click==7.1.2
email_validator
Flask==1.1.2
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
WTForms==2.3.3

dnspython~=1.16.0
idna~=2.10
requests~=2.24.0
pip~=20.2.3
lxml~=4.5.1
colorama~=0.4.3
Django~=3.0.5
python-dateutil~=2.8.1
setuptools~=47.1.0
pymongo~=3.10.1
certifi~=2020.4.5.1
rsa~=4.6
pyasn1~=0.4.8
parso~=0.7.0
jedi~=0.17.1
pytz~=2020.1
toml~=0.10.1
isort~=4.3.21
numpy~=1.19.0
networkx~=2.4
jsonpickle~=1.4.1
ipython~=7.16.1
six~=1.14.0
cachetools~=4.1.1
urllib3~=1.25.9
oauth2client~=4.1.3
pandas~=1.0.3
matplotlib~=3.2.2
xarray~=0.15.1
pylint~=2.5.3
astroid~=2.4.2
mccabe~=0.6.1
wrapt~=1.12.1
chardet~=3.0.4
Pygments~=2.6.1
ipykernel~=5.3.2
traitlets~=4.3.3
testpath~=0.4.4
decorator~=4.4.2
packaging~=20.4
backcall~=0.2.0
pickleshare~=0.7.5
wcwidth~=0.2.5
httplib2~=0.18.1
scipy~=1.4.1
oauthlib~=3.1.0
cycler~=0.10.0
tornado~=6.0.4
pyparsing~=2.4.7
kiwisolver~=1.2.0
pycodestyle~=2.6.0
mongoengine~=0.20.0
TexSoup~=0.2.1
gspread~=3.6.0
pyvis~=0.1.8.1.1
graphviz~=0.14
libsass~=0.20.1
dnspython<2.0.0,>=1.16.0; extra == "srv"
passlib~=1.7.4